Cada zumbi morto adiciona 0.5 ao life e cada gol tira 10 do life

+++ GameEntity.cpp
@@ -265,7 +265,7 @@ int Pawn::GetScore() const
 	return m_score;
 }
 
-void Pawn::AddToScore(const int score)
+void Pawn::AddToScore(const float score)
 {
 	m_score += score;
 }
