+++ PawnManager.h
@@ -17,13 +17,13 @@ class PawnManager
 	static std::vector<gs2d::math::Rect2Df> m_forbiddenAreas;
 
 public:
-	static const int GOAL_SCORE;
+	static const float GOAL_SCORE;
 
 	PawnManager(gs2d::VideoPtr video, BallPtr ball);
 	inline static const std::vector<gs2d::math::Rect2Df>& GetForbiddenAreas() { return m_forbiddenAreas; };
 	void Update(gs2d::VideoPtr video, gs2d::InputPtr input, gs2d::AudioPtr audio, EffectManagerPtr fxManager, const unsigned long lastFrameElapsedTimeMS, SpriteResourceManager& spr);
 	void Draw(gs2d::VideoPtr video, SpriteResourceManager& spr);
-	void AddToScore(const int pawn, const int score);
+	void AddToScore(const int pawn, const float score);
 };
 
 typedef boost::shared_ptr<PawnManager> PawnManagerPtr;
